requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
google-generativeai==0.8.3
python-dotenv==1.0.1
langgraph==0.2.60
langchain-google-genai==2.0.8
langchain-openai==0.2.14
langchain-core==0.3.28
aiosqlite==0.20.0
langgraph-checkpoint-sqlite==2.0.1
resend==2.19.0
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.32.3
jinja2==3.1.4
celery==5.3.6
redis==5.2.1
